Fix message duplication on flag update.
When updating e.g. the 'seen' flag of a message, it might happen that a
duplicate for that message is created. Root cause was that the wrong
local message for the remote message was put into the cache map, which
later led to message lookup failure, which made the code believe the
message still needs to be downloaded (again).

Additionally, ghost messages could be created after deleting messages
when the IDLE connection reported the 'deleted' flag being set after
doing the deletion locally. In that case, the local message was supposed
to not be present in the lookup map, but the remote message list was not
updated correctly to reflect the deletion.

Change-Id: I954ea65cfe17b11a7f9ba19020399ac8cc6c3745